Fix geo_scope on a model where that model acts_as_mappable :through => :some_column by automatically .include'ing the :through option
Also, ActsAsMappable::ClassMethods#geo_scope now returns an instance of
ActsAsMappable::Relation instead of ActiveRelation. Basically, we override
with the distance formula. In other words, if you say
Model.geo_scope(...).where("distance > 5") or
Model.geo_scope(...).order("distance desc"), that "distance" is automatically
replaced with the big long SQL that's generated (the Haversine formula, or
whatever). The downside is that you have to put the .where after the .geo_scope,
otherwise it won't work.