Fix find with :through
Basically, for a model which does this:

  acts_as_mappable :through => :some_column

queries using geo_scope automatically :include :some_column.

In a related vein, calling #geo_scope on a model now returns a special scope
object (ActsAsMappable::Relation instead of ActiveRecord::Relation) such that if
you call another scope method such as `where`, `order`, etc. and attempt to
refer to a "distance" column, that "distance" will automatically be replaced
with the Haversine formula in your query. No more HAVINGs or subqueries!!